Add matching for the smtp-server statement
One reason for this is that the statement can have two IP addresses, and
this gets confused for an IP and subnet mask.  There might be more of
these types of statements, so this might be later changed to match
multiple statements, all of which can have two or more IP addresses in
their syntax.