Match more than one IP subnet per line
This changes the routines to yield more than one IP subnet on a line
if found.

Also is able to match subnets consisting of an address and a separate
network mask (which also looks like an IP address).  It does this
using the proof-of-concept script parsetuples.py as the basis for the
mechanism used to parse over each line using finditer() and back up
in the string as necessary.