Adjust regexes to handle back references correctly
The two regexes EXTENDED_ACL_RE and EXTENDED_ASA_ACL_RE share
handle_extended_acl(), and this uses the same indexes for back
references for both.

In addition, due to the way objects in access-list statements are
referenced, there is some vagueness about what each object type is.
This can only be solved by testing the type of each object before
refrencing it.  That is, differentiating an object network from an
object service, and so on.