Add special cases to match Cisco's prefix-list behavior
Match Cisco's prefix-list behavior where "le" is set to 32 in cases
where it would have generated an error here.  That is, ordinarily if
"le" is not given, it defaults to the prefix length of the subnet.
If "ge" is a value greater than the length, this does not match
the invariant 0 <= ge <= le <= 32.  In this case, set "ge" to 32
instead.  This allows the following syntax, which is accepted by
Cisco prefix-lists: "10.0.0.0/8 ge 9".

The reverse situation where "ge" is not given is allowed here too,
but it is not allowed by Cisco.  E.g., "10.0.0.0/8 le 7".