Add two new options to make matching easier
Usually when searching for objects, it helps to take different
approaches such as trying the source addresses first, then doing
the same search on the destination addresses.  The -o (logical-or)
operation means you can do that in one operation.

To make it even easier, the -p (duplicate) option copies the same
address to the destination address.  This means that it is possible
to search for a subnet in either the source or destination using
something like "-o -p -s SUBNET".