Found better way of dealing with ACL header lines
When an ACL header or object is encountered, its name is saved so
that all subsequent network addresses, subnets, or objects are stored
under that name.  The problem is knowing when that name is no longer
relevant.  Under this new way of handling it, all lines that start
with a space are part of that name, and as soon as this is not true,
the name is no longer relevant.