Convert to use IPv4Network from ipaddress module
The ipaddress module has all of the functionality needed to replace
the home-spun routines that were here before (except in one situation
described below).  In particular, it has an overlaps() method that
replaces the previous containment test that is the heart of the IP
subnet matching routines.  Using this standard module will result in
more understandable code.

The one shortcoming is when parsing ACLs that use wildcard host masks,
such as those on Cisco routers.  Instead of 0.0.0.0/0.0.0.0 being
converted to 0.0.0.0/32, it becomes 0.0.0.0/0.  Hopefully this is
enough of a rare case that it won't matter much.