Add support for twice NAT statements
This attempts to match the source and destination objects correctly
for twice NAT statements as used on ASA firewalls.  The only exception
is a NAT statement using the address of the outgoing interface as
the global inside address.  The program is not yet able to cache the
interface addresses, so the inside address is changed to ANY instead.

As an example:

    interface Ethernet0/0
     nameif outside
     ip address 1.2.3.4 255.255.255.0
    !
    object network obj-10.1.1.1
      host 10.1.1.1
    object network obj-192.168.1.1
      host 192.168.1.1
    object network obj-172.16.1.1
      host 172.16.1.1
    !
    nat (inside,outside) source static obj-10.1.1.1 interface destination static obj-192.168.1.1 obj-172.16.1.1

In this case the NAT'd packet has the source and destination addresses
(1.2.3.4, 172.16.1.1), but it is currently parsed as (ANY, 172.16.1.1).