Change the return type of matching functions
The new return type for matching functions is a tuple.  The first
element is a string that specifies the type of the second element.
This makes it much clearer to see what the return type is, rather
than relying on testing the type of the argument returned.  This also
makes it easier to add new types.

The types are 'srcdst', 'position', 'header', and 'remark' (the
latter is not used yet but will be introduced in the next commit).
Most functions return 'srcdst' and the second element is another
tuple consisting of the source and destination networks from the
matching line.  Other types are used for miscellaneous operations.