Add options for dealing with line numbers
The new options are --line-number (-n), which enables printing
of absolute line numbers, and --acl-lines (-a), which prints the
relative line numbers for each ACL (used to show the line numbers
used for updating access-list statements).  The option -n changes
the behavior that line numbers are turned off by default.

This change required heavily reorganizing the parse_config() routine,
but now it feels better and is probably ready to be refactored.