Rename SrcDst to MatchObj
Rename all of the SrcDst named tuple objects to MatchObj.  The name of
the object better explains what its role is since it is now used for
more than source and destination networks.  This is in preparation
for adding a new named tuple also named SrcDst which contains the
source and destination networks in one object.