Experiment with using generators to keep state
Instead of keeping state machines in parse_config() for header tracking
and other things, break these out into separate functions using
coroutines (generators).  The way this works is that match objects
are sent individually to the coroutines, where they keep track of
the state internally.  If there is something that should be output
by parse_config(), an object with the information is sent back.

There are three coroutines: one for headers, one for ACL remarks,
and one for keeping track of the line number where an ACL starts.