Manage the coroutines in separate functions
The goal of this is to hide away the creation and message handling
of the three coroutines by using separate functions, gen_init()
and gen_send().  This makes parse_config() a little less cluttered,
but not as much as I was first thinking.  It is still to be determined
if this is worth it, which is why it is on a separate branch for now.

Also, provide type aliases for the generator functions so that the
typing is easier.  Even this has had limited usefulness as mypy
doesn't seem to handle some aspects of generator functions as well
as I had liked (see comment in code).