Remove the use of eval()
Instead of using eval() to get the value of the regular expression
string in MATCH_RE_TABLE_RAW, reorganize the table so that the match
functions are the key in the table.  This allows using __name__
to convert the function name into a str, and avoids eval().

A new dictionary, MATCH_FUNCS, uses the function name as a string to
map re_match.lastgroup into the function to call.