Avoid compiling match_re for each processed file
The match_re regex is used to parse the configuration.  Before the
change to process multiple files, it was compiled in the parse_config()
function.  This is the wrong place since it is called for each file.
Instead, move the regex compile up a level and pass the compiled
regex as an argument to parse_config().

The compilation is now done in prepare_regexes().  Combine this with
the creation of the MATCH_FUNCS dictionary.