Match the original IP address in network object NAT
Add more matching for network object NAT statements (auto NAT), this
time matching for the original IP address rather than the translated
IP address.  This turned out to be more difficult than first thought,
and required adding a new match type 'nat', a new coroutine, and
supporting code.

This also fixes a bug where the translated IP address was incorrectly
being added to the IP addresses for the object, and is the reason
for the new match type 'nat'.