Prevent object NAT from matching too much
Fix a problem with the previous change
<2022-07-29T11:55:36Z!julesreid@lavanauts.org> where it can match too
much when using the -d (destination) option alone.  This is because in
this case the source address matches any prefix and therefore matches
any object NAT object.  Only allow object NAT with original IP address
to match if the destination prefix allows any IP address or if the -r
(logical or) is used.