Fix bug in objectnat_coroutine()
The header_line state variable was not being reset correctly, and was
being held for longer than it should have.  It then made a recurrence
during parsing of twice NAT statements, far from the object NAT header
where the variable was first set.  Now it is reset on all passes
through the function except for the specific cases where it is needed.