Adjust SUBNET_RE to avoid matching invalid subnets
The previous regex allowed prefix lengths that were greater than 32,
which causes a NetmaskValueError when giving the string to IPv4Network.
Rather than ignore this error, adjust the regex to only match valid
prefix lengths.

Note that this can still accept what looks to be subnets with invalid
prefix lengths, such as "1.2.3.4/33", but what is really happening is
that the prefix length is not part of the match.  For this example,
it would only match as "1.2.3.4".

Also, to allow matching the format of a prefix length as an IP address
(such as "10.0.0.0/255.0.0.0"), adjust the SUBNET_WITH_MASK_RE regex
to allow a slash.  This is a valid format accepted by IPv4Network.