Handle multiple prefix ranges on command line
Multiple prefix ranges can be specified for the source or destination
arguments by separating each one with a comma.

A major reorganization was necessary, including adding a ParserState
class to handle the state for each permutation of the ranges, and
adding tracking of the lines that have already been printed in case
a line is matched more than once.

Updated README to document new feature.