Change -p option from duplicate to transpose
The -p option (duplicate the source address to match into the
destination address) was really only useful with -r (logical-OR).
This allows searching for an address appearing in the source or
desination of an access list.  Now that it is possible to match with
multiple parser states, the -p option can now instead be defined as
adding another simultaneous search with the source and destination
address swapped.  This provides a more holistic approach to the
options and can provide more powerful searches than before.

The -r (logical-OR) becomes unnecessary at this point because
simultaneous searches are done with an implicit logical OR between
them.  The option will be removed in the next commit.

The following table summarizes how the -x, -p, and -r options combine
with the -s and -d options.  The resulting parser states shows the
simultaneous searches that are done, with ∀ meaning "match anything".
The only search that can be done using -r which can't be done otherwise
is "-s m -d x", which results in "m,∀ ∀,x".  This doesn't seem
compelling enough to keep.

┃ option          │ source (-s) │ dest (-d) │ resulting (s,d) parser states   ┃
┣━━━━━━━━━━━━━━━━━┿━━━━━━━━━━━━━┿━━━━━━━━━━━┿━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━┫
┃                 │ m           │ -         │ m,∀                             ┃
┃                 │ m           │ x         │ m,x                             ┃
┃                 │ -           │ x         │ ∀,x                             ┃
┃                 │ -           │ -         │ ∀,∀                             ┃
┃ -x              │ -           │ -         │ ∀,∀                             ┃
┃ -x              │ -           │ x         │ x,∀                             ┃
┃ -x              │ m           │ x         │ x,m                             ┃
┃ -x              │ m           │ -         │ ∀,m                             ┃
┃ -p              │ m           │ -         │ m,∀ ∀,m                         ┃
┃ -p              │ m           │ x         │ m,x x,m                         ┃
┃ -p              │ -           │ x         │ ∀,x x,∀                         ┃
┃ -p              │ -           │ -         │ ∀,∀ ∀,∀                         ┃
┃ -r (deprecated) │ -           │ -         │ ∀,∀ ∀,∀                         ┃
┃ -r (deprecated) │ -           │ x         │ ∀,∀ ∀,x                         ┃
┃ -r (deprecated) │ m           │ x         │ m,∀ ∀,x                         ┃
┃ -r (deprecated) │ m           │ -         │ m,∀ ∀,∀                         ┃
┃                 │ m,n         │ -         │ m,∀ n,∀                         ┃
┃                 │ m,n         │ x,y       │ m,x m,y n,x n,y                 ┃
┃                 │ -           │ x,y       │ ∀,x ∀,y                         ┃
┃                 │ -           │ -         │ ∀,∀                             ┃
┃ -x              │ -           │ -         │ ∀,∀                             ┃
┃ -x              │ -           │ x,y       │ x,∀ y,∀                         ┃
┃ -x              │ m,n         │ x,y       │ x,m x,n y,m y,n                 ┃
┃ -x              │ m,n         │ -         │ ∀,m ∀,n                         ┃
┃ -p              │ m,n         │ -         │ m,∀ n,∀ ∀,m ∀,n                 ┃
┃ -p              │ m,n         │ x,y       │ m,x m,y n,x n,y x,m y,m x,n y,n ┃
┃ -p              │ -           │ x,y       │ ∀,x ∀,y x,∀ y,∀                 ┃
┃ -p              │ -           │ -         │ ∀,∀ ∀,∀                         ┃