Rewrite the matching code to be more flexible
The core of the matching code is now rewritten to be more flexible.
This is due to wanting to add a match coroutine for "description"
statements, but finding that the matching code had odd dependencies
that made it difficult to implement   The flexibility is that any
coroutine can always print a line by yielding it, and there is no
dependency on whether a line matched the search criteria.

A few of the coroutines had to make adjustments for this.  The code
to match "description" statements will come in the next commit.