Add a new match type 'notfound'
This helps delineate sections of the configuration that cannot be
detected without seeing all of the lines of the file.  Before, if
a line was not matched by any regular expression, it was not sent
to the state machines.  This caused problems for any state machine
like description_coroutine() which tests whether a line starts with
a space or not in order to know a new section is starting.