If not resolving a FQDN, use a subsitute address
When the -l option is turned off, FQDN names are unresolved.  For each
case, substitute the address 0.0.0.1/32 as a replacement so that the
entries can still be searched for if desired.  This address is in
the private block 0.0.0.0/8 ("this network"), but is a less-likely
address to appear in a configuration.

Also, if -l is turned on, return 0.0.0.2/32 if the gethostbyname_ex()
function returns an error.