Replace "interface" with IP address in tests
For the twice NAT tests, the inside global address was using the
keyword "interface", which is only partially supported (it uses
0.0.0.0/0 instead).  Replace it using an IP address from one of the
test networks TEST-NET-2 so that more tests can be done on the "nat"
statement.

Also, add comments to the objects within the data to emphasize what
role these IP addresses have.