Add requirement for bash
There is a requirement in the tests for an echo statement that
takes the -n argument.  This doesn't necessarily work on all Bourne
shells (such as on macOS).  Instead of trying to work around it,
the requirement is now for bash, which has this and other constructs
that might be used (such as using $'\t' to get a tab character).