Fix a bug in the parsing of header lines
In some cases, the state machine in header_coroutine() was not
detecting when a section closed (that is, not seeing when a group of
lines with space indentation changed to a line having no indentation).
This caused the wrong header line to be output when a matching line
was output.