Avoid ambiguity when using wildcard masks
Both regular subnet masks and wildcard masks are used by ACLs,
depending on the platform (routers tend to use wildcards, and ASAs
tend to use regular subnet masks).  While the ipaddress module handles
both types, there is ambiguity in one case: 0.0.0.0/0.0.0.0.  Is this
0.0.0.0/0 or 0.0.0.0/32?  The module returns the former.  Keep track
of what type of ACL is being used (i.e., whether the line is matched
from a router or from an ASA) to know which type should be returned.
It is possible to return the latter by removing the subnet mask.

>>> ipaddress.IPv4Network('0.0.0.0/0.0.0.0')
IPv4Network('0.0.0.0/0')
>>> ipaddress.IPv4Network('0.0.0.0')
IPv4Network('0.0.0.0/32')