Generate MATCH_RE_TABLE_RAW automatically
Before, this table was manually maintained to map the function which
handles the regex to the regex string itself.  Now, the decorator
handles converting the function name into the corresponding regex
string.  The regex string must be defined with an upper case identifier
which matches the function name in lower case.